% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_epv.R
\name{calc_epv}
\alias{calc_epv}
\title{Calculate effective tax rate}
\usage{
calc_epv(balance_df, income_df, cash_df, interest_rate = 0.03,
  discount_rate = 0.112)
}
\description{
Calculate effective tax rate
}
