% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_avi.R
\name{get_avi}
\alias{get_avi}
\title{Putting it altogether to retrieve Applied Value Investing Metrics}
\usage{
get_avi(symbol)
}
\description{
Putting it altogether to retrieve Applied Value Investing Metrics
}
