% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bond_riskf.R
\name{get_bond_riskf}
\alias{get_bond_riskf}
\title{Retrieve long term interest rate}
\usage{
get_bond_riskf(api_key = "DTynS_LtABNwms7mxX_y",
  treasury = "LT Composite > 10 Yrs")
}
\description{
importFrom Quandl Quandl
}
