% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen.R
\name{screen}
\alias{screen}
\title{Screen for stocks}
\usage{
screen(market_cap = list(), pe_ratio = list(), dividend_yield = list(),
  price_change_52week = list(), price_change_26week = list(),
  price_change_13week = list(), price_to_book = list(),
  net_profit_margin_percent_trailing_12months = list(),
  current_assets_to_liabilities_ratio_year = list(),
  longterm_debt_to_equity_year = list(), return_on_equity_5years = list(),
  net_income_growth_rate_5years = list(),
  revenue_growth_rate_10years = list(), eps_growth_rate_10years = list(),
  interest_coverage_year = list(), max_results = 1300)
}
\description{
Screen for stocks
}
